py-test/requirements.txt
pyserial==3.5
Pillow==8.4.0
opencv-python==4.5.4.60
